feat: Move internal/hmac to pkg/
This makes the hmac implementation used by go-scm public for use in other projects.
