+++ pkg/hmac/hmac.go
