+++ scm/driver/gitea/webhook.go
@@ -12,8 +12,8 @@ import (
 	"io/ioutil"
 	"net/http"
 
+	"github.com/jenkins-x/go-scm/pkg/hmac"
 	"github.com/jenkins-x/go-scm/scm"
-	"github.com/jenkins-x/go-scm/scm/driver/internal/hmac"
 )
 
 type webhookService struct {
